[all] Rename Precompiled to Common/Instantiate
