+++ Ponca/src/Common/instantiate/instantiate_list/entrypoint.h
@@ -5,4 +5,4 @@
 */
 
 // This file is just for having a better name to include
-#include "../../src/Common/instantiate/types.h"
+#include "types.h"
