--- Ponca/src/Precompiled/precompiled.cpp
@@ -1,1 +1,0 @@
-#include "precompiled.h"
