+++ examples/cpp/ponca_basic_cpu.cpp
@@ -14,7 +14,8 @@
 #include <iostream>
 #include <vector>
 
-#include <Ponca/Precompiled>
+#include <Ponca/Instantiate>
+#include <Ponca/Ponca>
 
 using namespace std;
 using namespace Ponca;
