+++ examples/cpp/ponca_binding.cpp
@@ -12,7 +12,8 @@
 
 #include <iostream>
 
-#include <Ponca/Precompiled>
+#include <Ponca/Instantiate>
+#include <Ponca/Ponca>
 
 using namespace std;
 using namespace Ponca;
