+++ examples/cpp/ponca_neighbor_search.cpp
@@ -11,7 +11,8 @@ file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #include <iostream>
 #include <random>
-#include <Ponca/Precompiled>
+#include <Ponca/Instantiate>
+#include <Ponca/Ponca>
 
 using DataPoint = Ponca::PointPosition<float, 3>;
