+++ examples/cuda/CMakeLists.txt
@@ -22,13 +22,13 @@ if( CMAKE_CUDA_COMPILER )
     add_dependencies(ponca-examples cuda_fit_plane)
     set_property(TARGET cuda_fit_plane PROPERTY CUDA_ARCHITECTURES OFF)
     target_compile_options(cuda_fit_plane PRIVATE $<$<COMPILE_LANGUAGE:CUDA>:--expt-relaxed-constexpr>)
-    target_link_libraries(cuda_fit_plane PUBLIC Ponca::Precompiled Eigen3::Eigen)
+    target_link_libraries(cuda_fit_plane PUBLIC Ponca::Instantiate Eigen3::Eigen)
 
     add_executable(cuda_fit_plane_kdtree "ponca_fit_kdtree.cu")
     add_dependencies(ponca-examples cuda_fit_plane_kdtree)
     set_property(TARGET cuda_fit_plane_kdtree PROPERTY CUDA_ARCHITECTURES OFF)
     target_compile_options(cuda_fit_plane_kdtree PRIVATE $<$<COMPILE_LANGUAGE:CUDA>:--expt-relaxed-constexpr>)
-    target_link_libraries(cuda_fit_plane_kdtree PUBLIC Ponca::Precompiled Eigen3::Eigen)
+    target_link_libraries(cuda_fit_plane_kdtree PUBLIC Ponca::Instantiate Eigen3::Eigen)
 
     add_subdirectory(ssgls)
 else()
