+++ tests/src/curvature_plane.cpp
@@ -18,7 +18,8 @@
 
 #include <vector>
 
-#include <Ponca/Precompiled>
+#include <Ponca/Instantiate>
+#include <Ponca/Ponca>
 #include <Ponca/SpatialPartitioning>
 
 using namespace std;
