+++ tests/src/fit_monge_patch.cpp
@@ -14,7 +14,8 @@
 #include "../common/testing.h"
 #include "../common/testUtils.h"
 
-#include <Ponca/Precompiled>
+#include <Ponca/Instantiate>
+#include <Ponca/Ponca>
 
 #include <vector>
 #include <fstream>
