+++ tests/src/fit_unoriented.cpp
@@ -12,7 +12,8 @@
 #include "../common/testing.h"
 #include "../common/testUtils.h"
 
-#include <Ponca/Precompiled>
+#include <Ponca/Instantiate>
+#include <Ponca/Ponca>
 
 #include <vector>
