+++ tests/src/projection.cpp
@@ -14,7 +14,8 @@
 
 #include "../common/testing.h"
 
-#include <Ponca/Precompiled>
+#include <Ponca/Instantiate>
+#include <Ponca/Ponca>
 
 #include <chrono>
 #include <math.h>
